Added get_item_by_id
